requirements.txt
anyio==4.4.0
ariadne==0.23.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.0.3
graphql-core==3.2.3
grpcio==1.66.1
grpcio-tools==1.66.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
protobuf==5.28.0
py-mon==2.0.5
requests==2.32.3
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.2
watchdog==5.0.0
Werkzeug==3.0.4
flask-cors==5.0.0
